Remove unused fields in saved index and avoid storing compressed vectors meta if empty (#7)

+++ search-worker/index/impl/io_macros.h
@@ -8,6 +8,8 @@
 #ifndef HAKES_SEARCHWORKER_INDEX_IMPL_IOMACRO_H_
 #define HAKES_SEARCHWORKER_INDEX_IMPL_IOMACRO_H_
 
+#include <cstdint>
+
 /*************************************************************
  * I/O macros
  *
@@ -18,7 +20,7 @@
 
 #define READANDCHECK(ptr, n)                         \
     {                                                \
-        size_t ret = (*f)(ptr, sizeof(*(ptr)), n);   \
+        uint64_t ret = (*f)(ptr, sizeof(*(ptr)), n);   \
         assert(ret == (n));                          \
     }
 
@@ -27,7 +29,7 @@
 // will fail if we write 256G of data at once...
 #define READVECTOR(vec)                                              \
     {                                                                \
-        size_t size;                                                 \
+        uint64_t size;                                                 \
         READANDCHECK(&size, 1);                                      \
         assert(size >= 0 && size < (uint64_t{1} << 40)); \
         (vec).resize(size);                                          \
@@ -36,14 +38,14 @@
 
 #define READSTRING(s)                     \
     {                                     \
-        size_t size = (s).size();         \
+        uint64_t size = (s).size();         \
         WRITEANDCHECK(&size, 1);          \
         WRITEANDCHECK((s).c_str(), size); \
     }
 
 #define WRITEANDCHECK(ptr, n)                         \
     {                                                 \
-        size_t ret = (*f)(ptr, sizeof(*(ptr)), n);    \
+        uint64_t ret = (*f)(ptr, sizeof(*(ptr)), n);    \
         assert(ret == (n));                           \
     }
 
@@ -51,7 +53,7 @@
 
 #define WRITEVECTOR(vec)                   \
     {                                      \
-        size_t size = (vec).size();        \
+        uint64_t size = (vec).size();        \
         WRITEANDCHECK(&size, 1);           \
         WRITEANDCHECK((vec).data(), size); \
     }
@@ -61,14 +63,14 @@
 #define WRITEXBVECTOR(vec)                         \
     {                                              \
         assert((vec).size() % 4 == 0); \
-        size_t size = (vec).size() / 4;            \
+        uint64_t size = (vec).size() / 4;            \
         WRITEANDCHECK(&size, 1);                   \
         WRITEANDCHECK((vec).data(), size * 4);     \
     }
 
 #define READXBVECTOR(vec)                                            \
     {                                                                \
-        size_t size;                                                 \
+        uint64_t size;                                                 \
         READANDCHECK(&size, 1);                                      \
         assert(size >= 0 && size < (uint64_t{1} << 40)); \
         size *= 4;                                                   \
